Restructuring program, breaking the monolith into multiple modules
