+++ anchor/programs/supply_chain/src/instructions/mod.rs
@@ -1,0 +1,6 @@
+pub use initialize_product::*;
+pub mod initialize_product;
+pub use log_event::*;
+pub mod log_event;
+pub use transfer_ownership::*;
+pub mod transfer_ownership;
