+++ anchor/programs/supply_chain/src/state/mod.rs
@@ -1,0 +1,4 @@
+pub use product::*;
+pub mod product;
+pub use event::*;
+pub mod event;
